requirements.txt
flask
yt-dlp
PyJWT
python-dotenv
redis
flask-socketio
gunicorn
eventlet